[refactor] Move politeiad api under correct dir.
Since we introduced the www backend the location of cmd and api were
confusing at best.  Move it into the correct subdir and cope.
