+++ politeiad/api/v1/identity/identity_test.go
