+++ politeiad/cmd/politeia/politeia.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/decred/dcrtime/merkle"
 	"github.com/decred/dcrutil"
-	"github.com/decred/politeia/api/v1"
-	"github.com/decred/politeia/api/v1/identity"
+	"github.com/decred/politeia/politeiad/api/v1"
+	"github.com/decred/politeia/politeiad/api/v1/identity"
 	"github.com/decred/politeia/util"
 	"github.com/kennygrant/sanitize"
 )
